Add more space in navbar to scroll top
