+++ components/Layout/Navigation/UserMenu.tsx
@@ -77,7 +77,10 @@ export default function InviteButton({ styles }) {
   preload(url, fetcher);
 
   const [anchorEl, setAnchorEl] = useState(null);
-  const handleClick = (e) => setAnchorEl(e.target);
+  const handleClick = (e) => {
+    e.stopPropagation();
+    setAnchorEl(e.target);
+  };
   const handleClose = () => setAnchorEl(null);
 
   return (
@@ -86,6 +89,7 @@ export default function InviteButton({ styles }) {
         anchorEl={anchorEl}
         open={Boolean(anchorEl)}
         onClose={handleClose}
+        onClick={(e) => e.stopPropagation()}
         transformOrigin={{
           vertical: "bottom",
           horizontal: "left",
