+++ components/Settings/Menus/LoginActivity.tsx
@@ -6,17 +6,17 @@ import {
   IconButton,
   ListItem,
   ListItemText,
-  Tooltip,
+  Tooltip
 } from "@mui/material";
 import { red } from "@mui/material/colors";
 import dayjs from "dayjs";
 import React from "react";
 import { Virtuoso } from "react-virtuoso";
 import { mutate } from "swr";
-import { fetchRawApi, useApi } from "../../lib/client/useApi";
-import { useSession } from "../../lib/client/useSession";
-import { ConfirmationModal } from "../ConfirmationModal";
-import { ErrorHandler } from "../Error";
+import { fetchRawApi, useApi } from "../../../lib/client/useApi";
+import { useSession } from "../../../lib/client/useSession";
+import { ConfirmationModal } from "../../ConfirmationModal";
+import { ErrorHandler } from "../../Error";
 
 const Session: any = React.memo(function Session({
   mutationUrl,
